fix(tun): handle invalid descriptors after reboot
